Clean up and refactor web auth
