+++ tests/unittests/general/web_middleware_test.py
@@ -2,10 +2,10 @@ from mock import patch
 
 from django.test import RequestFactory
 
-from nav.django.auth import ACCOUNT_ID_VAR
-from nav.django.auth import SUDOER_ID_VAR
-from nav.django.auth import AuthenticationMiddleware
-from nav.django import auth
+from nav.web.auth import ACCOUNT_ID_VAR
+from nav.web.auth import SUDOER_ID_VAR
+from nav.web.auth import AuthenticationMiddleware
+from nav.web import auth
 
 
 PLAIN_ACCOUNT = auth.Account(id=101, login='tim', password='wizard', locked=False)
@@ -20,7 +20,7 @@ class TestAuthenticationMiddleware(object):
         r = RequestFactory()
         fake_request = r.get('/')
         fake_request.session = {ACCOUNT_ID_VAR: PLAIN_ACCOUNT.id}
-        with patch('nav.django.auth.Account.objects.get', return_value=PLAIN_ACCOUNT):
+        with patch('nav.web.auth.Account.objects.get', return_value=PLAIN_ACCOUNT):
             AuthenticationMiddleware().process_request(fake_request)
             assert fake_request.account == PLAIN_ACCOUNT
             assert fake_request.session[ACCOUNT_ID_VAR] == fake_request.account.id
@@ -32,8 +32,8 @@ class TestAuthenticationMiddleware(object):
             ACCOUNT_ID_VAR: 101,
             SUDOER_ID_VAR: True,
         }
-        with patch('nav.django.auth.Account.objects.get', return_value=PLAIN_ACCOUNT):
-            with patch('nav.django.auth.get_sudoer', return_value='foo'):
+        with patch('nav.web.auth.Account.objects.get', return_value=PLAIN_ACCOUNT):
+            with patch('nav.web.auth.get_sudoer', return_value='foo'):
                 AuthenticationMiddleware().process_request(fake_request)
                 assert getattr(fake_request.account, 'sudo_operator', None) == 'foo'
 
@@ -41,7 +41,7 @@ class TestAuthenticationMiddleware(object):
         r = RequestFactory()
         fake_request = r.get('/')
         fake_request.session = {}
-        with patch('nav.django.auth.Account.objects.get', return_value=DEFAULT_ACCOUNT):
+        with patch('nav.web.auth.Account.objects.get', return_value=DEFAULT_ACCOUNT):
             AuthenticationMiddleware().process_request(fake_request)
             assert fake_request.account == DEFAULT_ACCOUNT
             assert fake_request.session[ACCOUNT_ID_VAR] == fake_request.account.id
